refactor(tests): consolidate CodeEditor tests, remove CSS-testing files
Delete 5 test files that test Tailwind class names and inline styles
rather than behavior (CodeEditor-layout-structure, CodeEditor-responsive,
CodeEditor-output-position, CodeEditor-markdown-mobile, CodeEditor-empty-states).

Consolidate 8 behavioral CodeEditor test files into a single
CodeEditor.test.tsx with one shared Monaco Editor mock declaration,
eliminating ~260 LOC of duplicated boilerplate. Remove toHaveClass/
toHaveStyle assertions from the consolidated file. Net ~1,800 LOC reduction.
